prefetch num microbatches
