+++ nemo/collections/nlp/models/nlp_model.py
@@ -14,7 +14,6 @@
 
 import copy
 import hashlib
-import itertools
 import json
 import os
 from typing import Any, Mapping, Optional
@@ -135,18 +134,6 @@ class NLPModel(ModelPT, Exportable):
             # register encoder config
             self.register_bert_model()
 
-    def _prefetch(self, iterator):
-        """Checks if the iterator still has elements to return.
-        Used in models using dataloader_iter to prefetch the next batch before fwd_bwd func
-        is called to avoid PP rank 2 from wait indefinitely to get outpits from PP 1
-        """
-        try:
-            element = next(iterator)
-        except StopIteration:
-            return iterator, True
-
-        # return a new iterator with the prefetched element reinserted at the front
-        return itertools.chain([element], iterator), False
 
     def register_artifact(
         self, config_path: str, src: str, verify_src_exists: bool = False,
